sicp: 2.1.2 Abstraction Barriers - rectangle representation - part 1
Implement a representation for rectangles in a plane. Then define procedures for
area and perimeter. Design the abstraction barriers, so that the same area and
perimeter procs can be reused even if the representation changes.

note: Commit the changes so that it's easier to see what has changed in part 2