sicp: 2.5.3 Example: Symbolic Algebra - dense polynomials (part 2.1)
I tried to implement dense and sparse polynomial packages as well as polynomial
package. I had dense-polynomial, sparse-polynomial and polynomial types and this
worked okay until I got to the point where I wanted to support (add sparse
dense). In order to achieve this I have to move the sparse/dense generic
operations on term-list level.

Here the operations are on a polynomial level and this won't work.