sicp: 2.5.3 Example: Symbolic Algebra: add/mul polys in diff vars
Add support for addition and multiplication of numbers and polynomials.
`(add/mul '(polynomial x (5 3) (3 2) (0 3)) 5)` will work now.

This is a prerequisite which will help me implement addition and multiplication
of polinomials in different variables.

I had to define add/mul for each pair of <poly,number-type>, because polynomial
is not part of the numbers tower type and I can't coerce a number to polynomial.